requirements.txt
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bleach==6.1.0
bqplot==0.12.42
branca==0.7.0
census==0.8.20
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colour==0.1.5
comm==0.2.0
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.7
django-cors-headers==4.3.0
django-rest-api==0.1.5
djangorestframework==3.14.0
duckdb==0.9.2
exceptiongroup==1.1.3
executing==2.0.1
fastjsonschema==2.18.1
filelock==3.13.1
fiona==1.9.5
folium==0.15.0
fonttools==4.45.1
gdown==4.7.1
geojson==3.1.0
geopandas==0.13.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipykernel==6.26.0
ipyleaflet==0.18.0
ipython==8.12.3
ipytree==0.2.2
ipywidgets==8.1.1
jedi==0.19.1
jellyfish==0.11.2
Jinja2==3.1.2
json5==0.9.14
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-client==8.6.0
jupyter-core==5.5.0
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter-server==2.10.0
jupyter-server-terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-server==2.25.0
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
leafmap==0.29.3
MarkupSafe==2.1.3
matplotlib==3.7.4
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook-shim==0.2.3
numpy==1.24.4
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pkgutil-resolve-name==1.3.10
platformdirs==3.11.0
plotly==5.18.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
psutil==5.9.6
psycopg==3.1.12
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
pyparsing==3.1.1
pyproj==3.5.0
pyshp==2.3.1
pystac==1.8.4
pystac-client==0.7.5
python-box==7.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.12.0
scooby==0.9.2
Send2Trash==1.8.2
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
tenacity==8.2.3
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
traittypes==0.2.1
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
us==3.1.1
wcwidth==0.2.9
webencodings==0.5.1
websocket-client==1.6.4
whitebox==2.3.1
whiteboxgui==2.3.0
widgetsnbextension==4.0.9
xyzservices==2023.10.1
zipp==3.17.0
